@@ -1,37 +1,57 @@
 #ifndef art_Framework_Core_ModuleMacros_h
 #define art_Framework_Core_ModuleMacros_h
 
-// ======================================================================
-//
+////////////////////////////////////////////////////////////////////////
 // ModuleMacros
 //
-// Note: Libraries that include these symbol definitions cannot be linked
-// into a main program as other libraries are.  This is because the "one
-// definition" rule would be violated.
+// Defines the macro DEFINE_ART_MODULE(<module_classname>) to be used in
+// XXX_module.cc to declare art modules (analyzers, filters, producers
+// and outputs).
 //
-// ======================================================================
+// Note: Libraries that include these symbol definitions cannot be
+// linked into a main program as other libraries are.  This is because
+// the "one definition" rule would be violated.
+//
+////////////////////////////////////////////////////////////////////////
 
 #include "art/Framework/Principal/WorkerParams.h"
+#include "art/Framework/Core/ModuleType.h"
 #include "art/Framework/Core/WorkerT.h"
+#include "art/Framework/Core/detail/ModuleTypeDeducer.h"
 #include "art/Persistency/Provenance/ModuleDescription.h"
 #include "cpp0x/memory"
 #include "fhiclcpp/ParameterSet.h"
 
-// DEFINE_ART_MODULE_TEMP is a short-term expedient, to get an early
-// version of art working. It will be removed as soon as feasible.
-#define DEFINE_ART_MODULE_TEMP(klass) \
-extern "C" \
-art::Worker* make_temp(art::WorkerParams const& wp, art::ModuleDescription const& md) \
-{ return new klass::WorkerType(std::unique_ptr<klass::ModuleType>(new klass(*(wp.pset_))), md, wp); }
+#include "art/Framework/Core/EventObserver.h"
 
-// produce the function that is used to create a module instance.
-#define DEFINE_ART_MODULE(klass) \
-extern "C" \
-std::unique_ptr<klass::ModuleType> make(fhicl::ParameterSet const& ps) \
-{ return std::unique_ptr<klass::ModuleType>(new klass(ps)); } \
- DEFINE_ART_MODULE_TEMP(klass)
+#include <type_traits>
 
-// ======================================================================
+// Function typedefs (used in art).
+namespace art {
+  namespace detail {
+    typedef art::Worker* (WorkerMaker_t) (art::WorkerParams const&,
+                                          art::ModuleDescription const&);
+    typedef art::ModuleType (ModuleTypeFunc_t) ();
+  }
+}
+
+// Produce the injected functions
+#define DEFINE_ART_MODULE(klass)                                        \
+  extern "C" {                                                          \
+    art::Worker*                                                        \
+    make_worker(art::WorkerParams const& wp,                            \
+                art::ModuleDescription const& md)                       \
+    {                                                                   \
+      return new                                                        \
+        klass::WorkerType(std::unique_ptr<klass::ModuleType>            \
+                          (new klass(wp.pset_)), md, wp);               \
+    }                                                                   \
+    art::ModuleType                                                     \
+    moduleType()                                                        \
+    {                                                                   \
+      return art::detail::ModuleTypeDeducer<klass::ModuleType>::value;  \
+    }                                                                   \
+  }
 
 #endif /* art_Framework_Core_ModuleMacros_h */
 
