@@ -14,8 +14,8 @@ using namespace std;
 
 namespace art {
   Path::Path(int bitpos, string const& path_name,
-             WorkersInPath const& workers,
-             TrigResPtr trptr,
+             WorkersInPath workers,
+             TrigResPtr pathResults,
              ParameterSet const&,
              ActionTable& actions,
              std::shared_ptr<ActivityRegistry> areg,
@@ -28,10 +28,10 @@ namespace art {
     state_(art::hlt::Ready),
     bitpos_(bitpos),
     name_(path_name),
-    trptr_(trptr),
+    trptr_(std::move(pathResults)),
     actReg_(areg),
     act_table_(&actions),
-    workers_(workers),
+    workers_(std::move(workers)),
     isEndPath_(isEndPath)
   {
   }
@@ -79,7 +79,7 @@ namespace art {
 
   void
   Path::recordStatus(int nwrwue, bool isEvent) {
-    if(isEvent) {
+    if (isEvent && trptr_) {
       (*trptr_)[bitpos_]=HLTPathStatus(state_, nwrwue);
     }
   }
