@@ -8,15 +8,12 @@
 using art::TriggerResultInserter;
 using fhicl::ParameterSet;
 
-TriggerResultInserter::TriggerResultInserter(const ParameterSet& pset, const TrigResPtr& trptr) :
-  trptr_(trptr),
+TriggerResultInserter::TriggerResultInserter(const ParameterSet& pset, HLTGlobalStatus & pathResults) :
+  trptr_(&pathResults),
   pset_id_(pset.id())
 {
   produces<TriggerResults>();
 }
-
-TriggerResultInserter::~TriggerResultInserter()
-{ }
 
 void TriggerResultInserter::produce(art::Event& e)
 {
