@@ -17,7 +17,7 @@ and manipulate Unix-style signal handling.
 namespace art {
 
     extern boost::mutex usr2_lock;
-    extern volatile bool shutdown_flag;
+    extern volatile int shutdown_flag;
 
     extern "C" {
       void ep_sigusr2(int,siginfo_t*,void*);
