@@ -237,7 +237,7 @@ int main()
   ServiceToken serviceToken_ =
     ServiceRegistry::createSet(ServiceRegistry::ParameterSets(), aReg);
 
-  serviceToken_.add(std::unique_ptr<TNS>(new TNS(proc_pset)));
+  serviceToken_.add(std::unique_ptr<TNS>(new TNS(proc_pset, paths)));
 
   //make the services available
   ServiceRegistry::Operate operate(serviceToken_);
