@@ -5,7 +5,7 @@
 #include "art/Framework/Principal/CurrentProcessingContext.h"
 #include "art/Framework/Core/Frameworkfwd.h"
 #include "art/Persistency/Provenance/MasterProductRegistry.h"
-#include "art/Framework/Core/ModuleFactory.h"
+#include "art/Framework/Core/detail/ModuleFactory.h"
 #include "art/Framework/Principal/WorkerParams.h"
 #include "art/Framework/Core/WorkerT.h"
 #include "art/Utilities/GetPassID.h"
@@ -22,14 +22,15 @@ namespace
   ModuleDescription
     createModuleDescription(WorkerParams const &p)
   {
-    ParameterSet const& procParams = *p.procPset_;
-    ParameterSet const& conf = *p.pset_;
-    ModuleDescription md;
-    md.parameterSetID_ = conf.id();
-    md.moduleName_ = conf.get<std::string>("module_type");
-    md.moduleLabel_ = conf.get<std::string>("module_label");
-    md.processConfiguration_ = ProcessConfiguration(p.processName_, procParams.id(), p.releaseVersion_, p.passID_);
-    return md;
+    ParameterSet const& procParams = p.procPset_;
+    ParameterSet const& conf = p.pset_;
+    return ModuleDescription(conf.id(),
+                             conf.get<std::string>("module_type"),
+                             conf.get<std::string>("module_label"),
+                             ProcessConfiguration(p.processName_,
+                                                  procParams.id(),
+                                                  getReleaseVersion(),
+                                                  getPassID()));
   }
 } // namespace
 
@@ -44,6 +45,8 @@ public:
   void setUp(){}
   void tearDown(){}
   void maker2Test();
+
+  detail::ModuleFactory fact;
 };
 
 ///registration of the test so that the runner can find it
@@ -63,11 +66,11 @@ void testmaker2::maker2Test()
   art::ActionTable table;
 
   art::MasterProductRegistry preg;
-  art::WorkerParams params1(p1, p1, preg, table, "PROD", art::getReleaseVersion(), art::getPassID());
-  art::WorkerParams params2(p2, p2, preg, table, "PROD", art::getReleaseVersion(), art::getPassID());
+  art::WorkerParams params1(p1, p1, preg, table, "PROD");
+  art::WorkerParams params2(p2, p2, preg, table, "PROD");
 
-  std::unique_ptr<Worker> w1 = ModuleFactory::makeWorker(params1, createModuleDescription(params1));
-  std::unique_ptr<Worker> w2 = ModuleFactory::makeWorker(params2, createModuleDescription(params2));
+  std::unique_ptr<Worker> w1 = fact.makeWorker(params1, createModuleDescription(params1));
+  std::unique_ptr<Worker> w2 = fact.makeWorker(params2, createModuleDescription(params2));
 
 //  return 0;
 }
