@@ -38,16 +38,16 @@ namespace {
                         fhicl::ParameterSet const &pset,
                         art::TypeID const &producedType) {
     checkTypeAndWrapper(producedType);
-    art::ModuleDescription md;
-    md.parameterSetID_ = pset.id();
-    md.moduleName_ = "arttest::NOMOD";
-    md.moduleLabel_ = moduleLabel;
-    md.processConfiguration_ =
-      art::ProcessConfiguration(processName,
-                                fhicl::ParameterSet().id(),
-                                art::getReleaseVersion(),
-                                "");
-    return art::BranchDescription(art::TypeLabel(bt, producedType, instanceName), md);
+    return
+      art::BranchDescription(art::TypeLabel(bt, producedType, instanceName),
+                             art::ModuleDescription(pset.id(),
+                                                    "arttest::NOMOD",
+                                                    moduleLabel,
+                                                    art::ProcessConfiguration
+                                                    (processName,
+                                                     fhicl::ParameterSet().id(),
+                                                     art::getReleaseVersion(),
+                                                     "")));
   }
 }
 
