@@ -18,6 +18,19 @@ namespace {
     os << text;
   }
 
+  struct CallBackClass {
+    void
+    func(std::ostream & os, std::string const & text)
+      {
+        os << text;
+      }
+    void
+    cfunc(std::ostream & os, std::string const & text) const
+      {
+        os << text;
+      }
+  };
+
   size_t const nSchedules = 3;
   art::ScheduleID const sID { 1 };
 }
@@ -30,6 +43,28 @@ BOOST_AUTO_TEST_CASE(TestSignal2_t)
   std::string const test_text { "Test text" };
   boost::test_tools::output_test_stream os;
   BOOST_CHECK_NO_THROW(s.watch(sID, testCallback));
+  BOOST_CHECK_NO_THROW(s.invoke(sID, os, test_text));
+  BOOST_CHECK(os.is_equal(test_text));
+}
+
+BOOST_AUTO_TEST_CASE(TestSignal2_func_t)
+{
+  TestSignal2 s(nSchedules);
+  std::string const test_text { "Test text" };
+  boost::test_tools::output_test_stream os;
+  CallBackClass cbc;
+  BOOST_CHECK_NO_THROW(s.watch(sID, &CallBackClass::func, cbc));
+  BOOST_CHECK_NO_THROW(s.invoke(sID, os, test_text));
+  BOOST_CHECK(os.is_equal(test_text));
+}
+
+BOOST_AUTO_TEST_CASE(TestSignal2_cfunc_t)
+{
+  TestSignal2 s(nSchedules);
+  std::string const test_text { "Test text" };
+  boost::test_tools::output_test_stream os;
+  CallBackClass const cbc;
+  BOOST_CHECK_NO_THROW(s.watch(sID, &CallBackClass::cfunc, cbc));
   BOOST_CHECK_NO_THROW(s.invoke(sID, os, test_text));
   BOOST_CHECK(os.is_equal(test_text));
 }
@@ -54,10 +89,10 @@ BOOST_AUTO_TEST_CASE(TestSignal0_t)
   boost::test_tools::output_test_stream os;
   // 2012/02/13 CG This explicit reference to base clase below is
   // necessary because std::ref(os) fails due to a private typedef
-  // output_test_stream::result_type (in Boost <=1.49.0) screwing up
-  // std::ref's attempt to determine whether output_test_stream is a
-  // callable entity.
-  std::ostringstream & osr(os);
+  // output_test_stream::result_type (in Boost <=1.53.0 at least)
+  // screwing up std::ref's attempt to determine whether
+  // output_test_stream is a callable entity.
+  std::ostringstream & osr __attribute__((unused))(os);
   BOOST_CHECK_NO_THROW(s.watch(sID,
                                std::bind(testCallback,
                                          std::ref(osr),
