@@ -7,7 +7,7 @@
 // from art v0_06_02.
 ////////////////////////////////////////////////////////////////////////
 
-#include <boost/test/included/unit_test.hpp>
+#include "boost/test/included/unit_test.hpp"
 
 #include "art/Framework/Core/EDAnalyzer.h"
 #include "art/Framework/IO/ProductMix/MixContainerTypes.h"
