@@ -2,7 +2,7 @@
 #include "art/Framework/Core/ModuleMacros.h"
 #include "cetlib/container_algorithms.h"
 
-#include <boost/test/included/unit_test.hpp>
+#include "boost/test/included/unit_test.hpp"
 
 #include <iostream>
 #include <set>
